fix: use python3 instead of python, check if service file exists before trying to delete it

+++ deploy_production_as_service.sh
@@ -28,17 +28,20 @@ cd backend
 
 printf "\n\n\x1B[33mStarting migration.\x1B[0m\n\n\n"
 
-sudo -E env PATH="$PATH" python manage.py clearsessions
+sudo -E env PATH="$PATH" python3 manage.py clearsessions
 
-sudo -E env PATH="$PATH" python manage.py makemigrations
+sudo -E env PATH="$PATH" python3 manage.py makemigrations
 
-sudo -E env PATH="$PATH" python manage.py migrate
+sudo -E env PATH="$PATH" python3 manage.py migrate
 
 printf "\n\n\x1B[33mMigration finished.\x1B[0m\n\n\n"
 
 cd ..
 
-sudo rm -r /etc/systemd/system/smartplug.service
+if [ -f /etc/systemd/system/smartplug.service ]; then
+	sudo rm -r /etc/systemd/system/smartplug.service
+fi
+
 sudo cp ./smartplug.service /etc/systemd/system/smartplug.service
 
 sudo systemctl daemon-reload
